requirements.txt
flask
celery
requests
markdownify

Flask-Discord
pyopenssl
GitPython
Flask-WTF
gunicorn
pandas
pybig
boto3
